renames the cron package
